test(hero): created class that implements the hero repositories
Created class that implements the hero repositories in memory and used the class in the test file
